add html, fix flask

+++ about/tools.py
@@ -5,4 +5,4 @@ def cosine_similarity(a: np.ndarray, b: np.ndarray) -> float:
     """
     Compute the cosine similarity between two vectors.
     """
-    return np.dot(a, b) / (np.linalg.norm(a) * np.linalg.norm(b))
+    return float(np.dot(a, b) / (np.linalg.norm(a) * np.linalg.norm(b)))
